Moved CoinFlipSampler into a new file as no longer a "rough function"
